requirements.txt
websockets 
aiohttp
vosk
soundfile
numpy
librosa
opuslib
pydub
PyOgg
audioread
webm
llama-cpp-python